storage-cache update
